requirements.txt
# Hermetic requirements for RL Market Maker
# Pinned versions for stability

# Core ML libraries
torch==2.1.1
numpy==1.24.3
gymnasium==0.29.1

# Data handling
pandas==2.0.3
pyyaml==6.0.1
tardis-dev==2.1.2
pyarrow==22.0.0

# Utilities
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
jupyter==1.0.0
ipykernel==6.25.0